passage sur le schéma de bdd de prod

+++ MegaCasting2022.DBLib/tables/Artiste.cs
@@ -5,17 +5,13 @@ namespace MegaCasting2022.DBLib.tables;
 
 public partial class Artiste : Utilisateur
 {
-    //public long Identifiant { get; set; }
-
     public string Cv { get; set; } = null!;
 
     public string Nom { get; set; } = null!;
 
     public string Prenom { get; set; } = null!;
 
-    public long IdentifiantSexe { get; set; }
-
-   // public virtual Utilisateur IdentifiantNavigation { get; set; } = null!;
+    public int IdentifiantSexe { get; set; }
 
     public virtual Sexe IdentifiantSexeNavigation { get; set; } = null!;
 }
