+++ MegaCasting2022.DBLib/tables/ContenuEditorial.cs
@@ -5,13 +5,7 @@ namespace MegaCasting2022.DBLib.tables;
 
 public partial class ContenuEditorial
 {
-    public long Identifiant { get; set; }
+    public int Identifiant { get; set; }
 
     public string Titre { get; set; } = null!;
-
-    //public virtual Conseil? Conseil { get; set; }
-
-    //public virtual FicheMetier? FicheMetier { get; set; }
-
-    //public virtual Interview? Interview { get; set; }
 }
